v2/requirements.txt
pyinstaller
PySide6-Fluent-Widgets[full]

